move audit to internal/pkg
