+++ cmd/dop/main.go
@@ -53,7 +53,7 @@ import (
 	_ "github.com/erda-project/erda/internal/apps/dop/providers/project/home"
 	_ "github.com/erda-project/erda/internal/apps/dop/providers/projectpipeline"
 	_ "github.com/erda-project/erda/internal/apps/dop/providers/taskerror"
-	_ "github.com/erda-project/erda/providers/audit"
+	_ "github.com/erda-project/erda/internal/pkg/audit"
 
 	// components
 	_ "github.com/erda-project/erda/internal/apps/dop/component-protocol/components"
