+++ cmd/openapi/main.go
@@ -23,6 +23,7 @@ import (
 	// providers and modules
 	_ "github.com/erda-project/erda-infra/providers"
 	_ "github.com/erda-project/erda-proto-go/core/token/client"
+	_ "github.com/erda-project/erda/internal/pkg/audit"
 	_ "github.com/erda-project/erda/internal/tools/openapi/legacy"
 	_ "github.com/erda-project/erda/internal/tools/openapi/openapi-ng"
 	_ "github.com/erda-project/erda/internal/tools/openapi/openapi-ng/auth"
@@ -44,7 +45,6 @@ import (
 	_ "github.com/erda-project/erda/internal/tools/openapi/openapi-ng/routes/dynamic/temporary"
 	_ "github.com/erda-project/erda/internal/tools/openapi/openapi-ng/routes/openapi-v1"
 	_ "github.com/erda-project/erda/internal/tools/openapi/openapi-ng/routes/proto"
-	_ "github.com/erda-project/erda/providers/audit"
 	_ "github.com/erda-project/erda/providers/service-discover/erda-discover"
 	_ "github.com/erda-project/erda/providers/service-discover/fixed-discover"
 )
