+++ internal/apps/msp/apm/notifygroup/notifygroup.service.go
@@ -25,9 +25,9 @@ import (
 	notifygroup "github.com/erda-project/erda-proto-go/core/messenger/notifygroup/pb"
 	"github.com/erda-project/erda-proto-go/msp/apm/notifygroup/pb"
 	tenantpb "github.com/erda-project/erda-proto-go/msp/tenant/pb"
+	"github.com/erda-project/erda/internal/pkg/audit"
 	"github.com/erda-project/erda/internal/tools/monitor/utils"
 	"github.com/erda-project/erda/pkg/common/errors"
-	"github.com/erda-project/erda/providers/audit"
 )
 
 type notifyGroupService struct {
