+++ internal/apps/msp/member/mebmer.service.go
@@ -25,9 +25,9 @@ import (
 	"github.com/erda-project/erda-proto-go/msp/member/pb"
 	projectpb "github.com/erda-project/erda-proto-go/msp/tenant/project/pb"
 	"github.com/erda-project/erda/apistructs"
+	"github.com/erda-project/erda/internal/pkg/audit"
 	"github.com/erda-project/erda/pkg/common/apis"
 	"github.com/erda-project/erda/pkg/common/errors"
-	"github.com/erda-project/erda/providers/audit"
 )
 
 type memberService struct {
