+++ internal/core/messenger/notifygroup/group.service.go
@@ -27,9 +27,9 @@ import (
 	"github.com/erda-project/erda/bundle"
 	"github.com/erda-project/erda/internal/core/legacy/services/notify"
 	"github.com/erda-project/erda/internal/core/legacy/services/permission"
+	"github.com/erda-project/erda/internal/pkg/audit"
 	"github.com/erda-project/erda/pkg/common/apis"
 	"github.com/erda-project/erda/pkg/common/errors"
-	"github.com/erda-project/erda/providers/audit"
 )
 
 type notifyGroupService struct {
