+++ internal/core/messenger/notifygroup/provider.go
@@ -28,9 +28,9 @@ import (
 	"github.com/erda-project/erda/internal/core/legacy/dao"
 	"github.com/erda-project/erda/internal/core/legacy/services/notify"
 	"github.com/erda-project/erda/internal/core/legacy/services/permission"
+	"github.com/erda-project/erda/internal/pkg/audit"
 	"github.com/erda-project/erda/pkg/common/apis"
 	"github.com/erda-project/erda/pkg/i18n"
-	"github.com/erda-project/erda/providers/audit"
 )
 
 type config struct{}
