+++ internal/tools/monitor/core/alert/alert-apis/alert.service.go
@@ -31,13 +31,13 @@ import (
 	"github.com/erda-project/erda-proto-go/core/monitor/alert/pb"
 	metricpb "github.com/erda-project/erda-proto-go/core/monitor/metric/pb"
 	"github.com/erda-project/erda/apistructs"
+	"github.com/erda-project/erda/internal/pkg/audit"
 	"github.com/erda-project/erda/internal/tools/monitor/core/alert/alert-apis/adapt"
 	"github.com/erda-project/erda/internal/tools/monitor/core/alert/alert-apis/db"
 	"github.com/erda-project/erda/internal/tools/monitor/utils"
 	"github.com/erda-project/erda/pkg/common/apis"
 	"github.com/erda-project/erda/pkg/common/errors"
 	api "github.com/erda-project/erda/pkg/common/httpapi"
-	"github.com/erda-project/erda/providers/audit"
 )
 
 type alertService struct {
