+++ internal/tools/monitor/core/dataview/data.view.service.go
@@ -26,10 +26,10 @@ import (
 
 	"github.com/erda-project/erda-proto-go/core/monitor/dataview/pb"
 	"github.com/erda-project/erda/cmd/monitor/monitor/conf"
+	"github.com/erda-project/erda/internal/pkg/audit"
 	"github.com/erda-project/erda/internal/tools/monitor/core/dataview/db"
 	"github.com/erda-project/erda/pkg/common/apis"
 	"github.com/erda-project/erda/pkg/common/errors"
-	"github.com/erda-project/erda/providers/audit"
 )
 
 type dataViewService struct {
